+++ src/app/cart/cart.module.ts
@@ -3,6 +3,7 @@ import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
 import { CartRoutingModule } from './cart-routing.module';
+import { localStorageInstance, LocalStorageService } from '../core/services/local-storage.service';
 
 @NgModule({
   declarations: [
@@ -15,6 +16,9 @@ import { CartRoutingModule } from './cart-routing.module';
   ],
   exports: [
     CartListComponent
+  ],
+  providers: [
+    { provide: LocalStorageService, useValue: localStorageInstance }
   ]
 })
 export class CartModule { }