+++ changelog.md
@@ -61,3 +61,4 @@ Added edit form and resolve guard
 03.03.2021
 Added edit/add possibility for admin; Added spinner
 Added CanDeactivate guard
+Added auth service and Roles enum; Added CanLoad and CanActivate guards to admin module